Refactored epic unit tests
