+++ raiden-ts/tests/unit/epics/channels.spec.ts
@@ -1,0 +1,1000 @@
+import { marbles } from 'rxjs-marbles/jest';
+import { of, from, timer } from 'rxjs';
+import { first, takeUntil, toArray, delay } from 'rxjs/operators';
+import { ContractTransaction } from 'ethers/contract';
+import { bigNumberify } from 'ethers/utils';
+import { Zero, HashZero } from 'ethers/constants';
+import { defaultAbiCoder } from 'ethers/utils/abi-coder';
+import { getType } from 'typesafe-actions';
+import { range } from 'lodash';
+
+import { UInt } from 'raiden-ts/utils/types';
+import { RaidenAction } from 'raiden-ts/actions';
+import { RaidenState } from 'raiden-ts/state';
+import {
+  newBlock,
+  tokenMonitored,
+  channelMonitored,
+  channelOpen,
+  channelOpened,
+  channelDeposit,
+  channelDeposited,
+  channelClose,
+  channelClosed,
+  channelSettleable,
+  channelSettle,
+  channelSettled,
+  channelOpenFailed,
+  channelDepositFailed,
+  channelCloseFailed,
+  channelSettleFailed,
+  channelWithdrawn,
+} from 'raiden-ts/channels/actions';
+import {
+  channelOpenEpic,
+  channelOpenedEpic,
+  channelDepositEpic,
+  channelCloseEpic,
+  channelSettleEpic,
+  channelMonitoredEpic,
+  channelSettleableEpic,
+} from 'raiden-ts/channels/epics';
+import { raidenReducer } from 'raiden-ts/reducer';
+
+import { epicFixtures } from '../fixtures';
+import { raidenEpicDeps, makeLog } from '../mocks';
+
+describe('channels epic', () => {
+  const depsMock = raidenEpicDeps();
+  const {
+    token,
+    tokenContract,
+    tokenNetworkContract,
+    tokenNetwork,
+    channelId,
+    partner,
+    settleTimeout,
+    isFirstParticipant,
+    txHash,
+    state,
+  } = epicFixtures(depsMock);
+
+  afterEach(() => {
+    jest.clearAllMocks();
+  });
+
+  test(
+    'channelSettleableEpic',
+    marbles(m => {
+      const closeBlock = 125;
+      // state contains one channel in closed state
+      const newState = [
+        tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+        channelOpened(
+          { id: channelId, settleTimeout, openBlock: 121, isFirstParticipant, txHash },
+          { tokenNetwork, partner },
+        ),
+        channelClosed(
+          {
+            id: channelId,
+            participant: depsMock.address,
+            closeBlock,
+            txHash,
+          },
+          { tokenNetwork, partner },
+        ),
+      ].reduce(raidenReducer, state);
+      /* first newBlock bigger than settleTimeout causes a channelSettleable to be emitted */
+      const action$ = m.cold('---b-B-|', {
+          b: newBlock({ blockNumber: closeBlock + settleTimeout - 1 }),
+          B: newBlock({ blockNumber: closeBlock + settleTimeout + 4 }),
+        }),
+        state$ = m.cold('--s-|', { s: newState });
+      m.expect(channelSettleableEpic(action$, state$)).toBeObservable(
+        m.cold('-----S-|', {
+          S: channelSettleable(
+            { settleableBlock: closeBlock + settleTimeout + 4 },
+            { tokenNetwork, partner },
+          ),
+        }),
+      );
+    }),
+  );
+
+  describe('channelOpenEpic', () => {
+    test('fails if channel.state !== opening', async () => {
+      // there's a channel already opened in state
+      const action = channelOpen({ settleTimeout }, { tokenNetwork, partner }),
+        curState = [
+          tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+          channelOpened(
+            { id: channelId, settleTimeout, openBlock: 125, isFirstParticipant, txHash },
+            { tokenNetwork, partner },
+          ),
+        ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(action),
+        state$ = of<RaidenState>(curState);
+
+      await expect(channelOpenEpic(action$, state$, depsMock).toPromise()).resolves.toMatchObject({
+        type: getType(channelOpenFailed),
+        payload: expect.any(Error),
+        error: true,
+        meta: { tokenNetwork, partner },
+      });
+    });
+
+    test('tx fails', async () => {
+      const action = channelOpen({ settleTimeout }, { tokenNetwork, partner }),
+        curState = [tokenMonitored({ token, tokenNetwork, fromBlock: 1 }), action].reduce(
+          raidenReducer,
+          state,
+        );
+      const action$ = of<RaidenAction>(action),
+        state$ = of<RaidenState>(curState);
+
+      const tx: ContractTransaction = {
+        hash: txHash,
+        confirmations: 1,
+        nonce: 1,
+        gasLimit: bigNumberify(1e6),
+        gasPrice: bigNumberify(2e10),
+        value: Zero,
+        data: '0x',
+        chainId: depsMock.network.chainId,
+        from: depsMock.address,
+        wait: jest.fn().mockResolvedValue({ byzantium: true, status: 0 }),
+      };
+      tokenNetworkContract.functions.openChannel.mockResolvedValueOnce(tx);
+
+      await expect(channelOpenEpic(action$, state$, depsMock).toPromise()).resolves.toMatchObject({
+        type: getType(channelOpenFailed),
+        payload: expect.any(Error),
+        error: true,
+        meta: { tokenNetwork, partner },
+      });
+    });
+
+    test('success', async () => {
+      // there's a channel already opened in state
+      const action = channelOpen({ settleTimeout }, { tokenNetwork, partner }),
+        curState = [tokenMonitored({ token, tokenNetwork, fromBlock: 1 }), action].reduce(
+          raidenReducer,
+          state,
+        );
+      const action$ = of<RaidenAction>(action),
+        state$ = of<RaidenState>(curState);
+
+      const tx: ContractTransaction = {
+        hash: txHash,
+        confirmations: 1,
+        nonce: 1,
+        gasLimit: bigNumberify(1e6),
+        gasPrice: bigNumberify(2e10),
+        value: Zero,
+        data: '0x',
+        chainId: depsMock.network.chainId,
+        from: depsMock.address,
+        wait: jest.fn().mockResolvedValue({ byzantium: true, status: 1 }),
+      };
+      tokenNetworkContract.functions.openChannel.mockResolvedValueOnce(tx);
+
+      // result is undefined on success as the respective channelOpenedAction is emitted by the
+      // tokenMonitoredEpic, which monitors the blockchain for ChannelOpened events
+      await expect(
+        channelOpenEpic(action$, state$, depsMock).toPromise(),
+      ).resolves.toBeUndefined();
+      expect(tokenNetworkContract.functions.openChannel).toHaveBeenCalledTimes(1);
+      expect(tx.wait).toHaveBeenCalledTimes(1);
+    });
+  });
+
+  describe('channelOpenedEpic', () => {
+    const {
+      token,
+      tokenNetwork,
+      channelId,
+      partner,
+      settleTimeout,
+      isFirstParticipant,
+      txHash,
+      state,
+    } = epicFixtures(depsMock);
+
+    afterEach(() => {
+      jest.clearAllMocks();
+    });
+
+    test("filter out if channel isn't in 'open' state", async () => {
+      // channel.state is 'opening'
+      const curState = [
+        tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+        channelOpen({ settleTimeout }, { tokenNetwork, partner }),
+      ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(
+          channelOpened(
+            { id: channelId, settleTimeout, openBlock: 125, isFirstParticipant, txHash },
+            { tokenNetwork, partner },
+          ),
+        ),
+        state$ = of<RaidenState>(curState);
+
+      await expect(channelOpenedEpic(action$, state$).toPromise()).resolves.toBeUndefined();
+    });
+
+    test('channelOpened triggers channel monitoring', async () => {
+      // channel.state is 'opening'
+      const action = channelOpened(
+          { id: channelId, settleTimeout, openBlock: 125, isFirstParticipant, txHash },
+          { tokenNetwork, partner },
+        ),
+        curState = [tokenMonitored({ token, tokenNetwork, fromBlock: 1 }), action].reduce(
+          raidenReducer,
+          state,
+        );
+      const action$ = of<RaidenAction>(action),
+        state$ = of<RaidenState>(curState);
+
+      await expect(channelOpenedEpic(action$, state$).toPromise()).resolves.toMatchObject({
+        type: getType(channelMonitored),
+        payload: { id: channelId, fromBlock: 125 },
+        meta: { tokenNetwork, partner },
+      });
+    });
+  });
+
+  describe('channelMonitoredEpic', () => {
+    const deposit = bigNumberify(1023) as UInt<32>,
+      depositEncoded = defaultAbiCoder.encode(['uint256'], [deposit]),
+      withdraw = bigNumberify(100) as UInt<32>,
+      withdrawEncoded = defaultAbiCoder.encode(['uint256'], [withdraw]),
+      openBlock = 121,
+      closeBlock = 124,
+      settleBlock = closeBlock + settleTimeout + 1,
+      settleDataEncoded = defaultAbiCoder.encode(
+        ['uint256', 'bytes32', 'uint256', 'bytes32'],
+        [Zero, HashZero, Zero, HashZero],
+      );
+
+    test('first channelMonitored with past$ own ChannelNewDeposit event', async () => {
+      const curState = [
+        tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+        channelOpened(
+          { id: channelId, settleTimeout, openBlock, isFirstParticipant, txHash },
+          { tokenNetwork, partner },
+        ),
+      ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(
+          channelMonitored({ id: channelId, fromBlock: openBlock }, { tokenNetwork, partner }),
+        ).pipe(delay(1)), // give time to state multicast to register
+        state$ = of<RaidenState>(curState);
+
+      depsMock.provider.getLogs.mockResolvedValueOnce([
+        makeLog({
+          blockNumber: 123,
+          filter: tokenNetworkContract.filters.ChannelNewDeposit(
+            channelId,
+            depsMock.address,
+            null,
+          ),
+          data: depositEncoded, // non-indexed total_deposit = 1023 goes in data
+        }),
+      ]);
+
+      await expect(
+        channelMonitoredEpic(action$, state$, depsMock)
+          .pipe(first())
+          .toPromise(),
+      ).resolves.toMatchObject({
+        type: getType(channelDeposited),
+        payload: { id: channelId, participant: depsMock.address, totalDeposit: deposit },
+        meta: { tokenNetwork, partner },
+      });
+    });
+
+    test('already channelMonitored with new$ partner ChannelNewDeposit event', async () => {
+      const action = channelMonitored({ id: channelId }, { tokenNetwork, partner }),
+        curState = [
+          tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+          channelOpened(
+            { id: channelId, settleTimeout, openBlock, isFirstParticipant, txHash },
+            { tokenNetwork, partner },
+          ),
+        ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(action),
+        state$ = of<RaidenState>(curState);
+
+      const promise = channelMonitoredEpic(action$, state$, depsMock)
+        .pipe(first())
+        .toPromise();
+
+      depsMock.provider.emit(
+        tokenNetworkContract.filters.ChannelNewDeposit(channelId, null, null),
+        makeLog({
+          blockNumber: 125,
+          filter: tokenNetworkContract.filters.ChannelNewDeposit(channelId, partner, null),
+          data: depositEncoded, // non-indexed total_deposit = 1023 goes in data
+        }),
+      );
+
+      await expect(promise).resolves.toMatchObject({
+        type: getType(channelDeposited),
+        payload: { id: channelId, participant: partner, totalDeposit: deposit },
+        meta: { tokenNetwork, partner },
+      });
+    });
+
+    test("ensure multiple channelMonitored don't produce duplicated events", async () => {
+      const multiple = 16;
+      const curState = [
+        tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+        channelOpened(
+          { id: channelId, settleTimeout, openBlock, isFirstParticipant, txHash },
+          { tokenNetwork, partner },
+        ),
+      ].reduce(raidenReducer, state);
+      const action$ = from(
+          range(multiple).map(() =>
+            channelMonitored({ id: channelId }, { tokenNetwork, partner }),
+          ),
+        ),
+        state$ = of<RaidenState>(curState);
+
+      const promise = channelMonitoredEpic(action$, state$, depsMock)
+        .pipe(
+          // wait a little and then complete observable, so it doesn't keep listening forever
+          takeUntil(timer(100)),
+          toArray(), // aggregate all emitted values in this period in a single array
+        )
+        .toPromise();
+
+      // even though multiple channelMonitored events were fired, blockchain fires a single event
+      depsMock.provider.emit(
+        tokenNetworkContract.filters.ChannelNewDeposit(channelId, null, null),
+        makeLog({
+          blockNumber: 125,
+          filter: tokenNetworkContract.filters.ChannelNewDeposit(
+            channelId,
+            depsMock.address,
+            null,
+          ),
+          data: depositEncoded, // non-indexed total_deposit = 1023 goes in data
+        }),
+      );
+
+      const result = await promise;
+      expect(result).toHaveLength(1);
+      expect(result[0]).toMatchObject({
+        type: getType(channelDeposited),
+        payload: { id: channelId, participant: depsMock.address, totalDeposit: deposit },
+        meta: { tokenNetwork, partner },
+      });
+
+      expect(depsMock.provider.on).toHaveBeenCalledTimes(4); // one for each event
+    });
+
+    test('new$ partner ChannelWithdraw event', async () => {
+      const curState = [
+        tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+        channelOpened(
+          { id: channelId, settleTimeout, openBlock, isFirstParticipant, txHash },
+          { tokenNetwork, partner },
+        ),
+        channelDeposited(
+          {
+            id: channelId,
+            participant: partner,
+            totalDeposit: bigNumberify(410) as UInt<32>,
+            txHash,
+          },
+          { tokenNetwork, partner },
+        ),
+      ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(
+          channelMonitored({ id: channelId }, { tokenNetwork, partner }),
+        ),
+        state$ = of<RaidenState>(curState);
+
+      const promise = channelMonitoredEpic(action$, state$, depsMock)
+        .pipe(first())
+        .toPromise();
+
+      depsMock.provider.emit(
+        tokenNetworkContract.filters.ChannelWithdraw(channelId, null, null),
+        makeLog({
+          blockNumber: closeBlock,
+          transactionHash: txHash,
+          filter: tokenNetworkContract.filters.ChannelWithdraw(channelId, partner, null),
+          data: withdrawEncoded, // non-indexed totalWithdraw
+        }),
+      );
+
+      await expect(promise).resolves.toMatchObject({
+        type: getType(channelWithdrawn),
+        payload: { id: channelId, participant: partner, totalWithdraw: withdraw, txHash },
+        meta: { tokenNetwork, partner },
+      });
+    });
+
+    test('new$ partner ChannelClosed event', async () => {
+      const curState = [
+        tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+        channelOpened(
+          { id: channelId, settleTimeout, openBlock, isFirstParticipant, txHash },
+          { tokenNetwork, partner },
+        ),
+      ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(
+          channelMonitored({ id: channelId }, { tokenNetwork, partner }),
+        ),
+        state$ = of<RaidenState>(curState);
+
+      const promise = channelMonitoredEpic(action$, state$, depsMock)
+        .pipe(first())
+        .toPromise();
+
+      depsMock.provider.emit(
+        tokenNetworkContract.filters.ChannelClosed(channelId, null, null, null),
+        makeLog({
+          blockNumber: closeBlock,
+          transactionHash: txHash,
+          filter: tokenNetworkContract.filters.ChannelClosed(channelId, partner, 11, null),
+          data: HashZero, // non-indexed balance_hash
+        }),
+      );
+
+      await expect(promise).resolves.toMatchObject({
+        type: getType(channelClosed),
+        payload: { id: channelId, participant: partner, closeBlock, txHash },
+        meta: { tokenNetwork, partner },
+      });
+    });
+
+    test('new$ ChannelSettled event', async () => {
+      const curState = [
+        tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+        channelOpened(
+          { id: channelId, settleTimeout, openBlock, isFirstParticipant, txHash },
+          { tokenNetwork, partner },
+        ),
+        channelClosed(
+          { id: channelId, participant: depsMock.address, closeBlock, txHash },
+          { tokenNetwork, partner },
+        ), // channel is in "closed" state already
+      ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(
+          channelMonitored({ id: channelId }, { tokenNetwork, partner }),
+        ),
+        state$ = of<RaidenState>(curState);
+
+      expect(depsMock.provider.removeListener).not.toHaveBeenCalled();
+      const promise = channelMonitoredEpic(action$, state$, depsMock)
+        .pipe(takeUntil(timer(100)))
+        .toPromise();
+
+      expect(
+        tokenNetworkContract.listenerCount(
+          tokenNetworkContract.filters.ChannelNewDeposit(channelId, null, null),
+        ),
+      ).toBe(1);
+
+      expect(
+        tokenNetworkContract.listenerCount(
+          tokenNetworkContract.filters.ChannelClosed(channelId, null, null, null),
+        ),
+      ).toBe(1);
+
+      expect(
+        tokenNetworkContract.listenerCount(
+          tokenNetworkContract.filters.ChannelSettled(channelId, null, null, null, null),
+        ),
+      ).toBe(1);
+
+      depsMock.provider.emit(
+        tokenNetworkContract.filters.ChannelSettled(channelId, null, null, null, null),
+        makeLog({
+          blockNumber: settleBlock,
+          transactionHash: txHash,
+          filter: tokenNetworkContract.filters.ChannelSettled(channelId, null, null, null, null),
+          data: settleDataEncoded, // participants amounts aren't indexed, so they go in data
+        }),
+      );
+
+      await expect(promise).resolves.toEqual(
+        channelSettled({ id: channelId, settleBlock, txHash }, { tokenNetwork, partner }),
+      );
+
+      // ensure ChannelSettledAction completed channel monitoring and unsubscribed from events
+      expect(depsMock.provider.removeListener).toHaveBeenCalledWith(
+        tokenNetworkContract.filters.ChannelNewDeposit(channelId, null, null),
+        expect.anything(),
+      );
+
+      expect(depsMock.provider.removeListener).toHaveBeenCalledWith(
+        tokenNetworkContract.filters.ChannelClosed(channelId, null, null, null),
+        expect.anything(),
+      );
+
+      expect(depsMock.provider.removeListener).toHaveBeenCalledWith(
+        tokenNetworkContract.filters.ChannelSettled(channelId, null, null, null, null),
+        expect.anything(),
+      );
+
+      expect(
+        tokenNetworkContract.listenerCount(
+          tokenNetworkContract.filters.ChannelNewDeposit(channelId, null, null),
+        ),
+      ).toBe(0);
+
+      expect(
+        tokenNetworkContract.listenerCount(
+          tokenNetworkContract.filters.ChannelClosed(channelId, null, null, null),
+        ),
+      ).toBe(0);
+
+      expect(
+        tokenNetworkContract.listenerCount(
+          tokenNetworkContract.filters.ChannelSettled(channelId, null, null, null, null),
+        ),
+      ).toBe(0);
+    });
+  });
+
+  describe('channelDepositEpic', () => {
+    const deposit = bigNumberify(1023) as UInt<32>,
+      openBlock = 121;
+
+    test('fails if there is no token for tokenNetwork', async () => {
+      // there's a channel already opened in state
+      const action$ = of<RaidenAction>(channelDeposit({ deposit }, { tokenNetwork, partner })),
+        state$ = of<RaidenState>(state);
+
+      await expect(
+        channelDepositEpic(action$, state$, depsMock).toPromise(),
+      ).resolves.toMatchObject({
+        type: getType(channelDepositFailed),
+        payload: expect.any(Error),
+        error: true,
+        meta: { tokenNetwork, partner },
+      });
+    });
+
+    test('fails if channel.state !== "open"', async () => {
+      // there's a channel already opened in state
+      const action = channelDeposit({ deposit }, { tokenNetwork, partner }),
+        // channel is in 'opening' state
+        curState = [
+          tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+          channelOpen({ settleTimeout }, { tokenNetwork, partner }),
+        ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(action),
+        state$ = of<RaidenState>(curState);
+
+      await expect(
+        channelDepositEpic(action$, state$, depsMock).toPromise(),
+      ).resolves.toMatchObject({
+        type: getType(channelDepositFailed),
+        payload: expect.any(Error),
+        error: true,
+        meta: { tokenNetwork, partner },
+      });
+    });
+
+    test('approve tx fails', async () => {
+      // there's a channel already opened in state
+      const curState = [
+        tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+        channelOpened(
+          { id: channelId, settleTimeout, openBlock, isFirstParticipant, txHash },
+          { tokenNetwork, partner },
+        ),
+      ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(channelDeposit({ deposit }, { tokenNetwork, partner })),
+        state$ = of<RaidenState>(curState);
+
+      const approveTx: ContractTransaction = {
+        hash: txHash,
+        confirmations: 1,
+        nonce: 1,
+        gasLimit: bigNumberify(1e6),
+        gasPrice: bigNumberify(2e10),
+        value: Zero,
+        data: '0x',
+        chainId: depsMock.network.chainId,
+        from: depsMock.address,
+        wait: jest.fn().mockResolvedValue({ byzantium: true, status: 0 }),
+      };
+      tokenContract.functions.approve.mockResolvedValueOnce(approveTx);
+
+      await expect(
+        channelDepositEpic(action$, state$, depsMock).toPromise(),
+      ).resolves.toMatchObject({
+        type: getType(channelDepositFailed),
+        payload: expect.any(Error),
+        error: true,
+        meta: { tokenNetwork, partner },
+      });
+    });
+
+    test('setTotalDeposit tx fails', async () => {
+      // there's a channel already opened in state
+      const curState = [
+        tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+        channelOpened(
+          { id: channelId, settleTimeout, openBlock, isFirstParticipant, txHash },
+          { tokenNetwork, partner },
+        ),
+      ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(channelDeposit({ deposit }, { tokenNetwork, partner })),
+        state$ = of<RaidenState>(curState);
+
+      const approveTx: ContractTransaction = {
+        hash: txHash,
+        confirmations: 1,
+        nonce: 1,
+        gasLimit: bigNumberify(1e6),
+        gasPrice: bigNumberify(2e10),
+        value: Zero,
+        data: '0x',
+        chainId: depsMock.network.chainId,
+        from: depsMock.address,
+        wait: jest.fn().mockResolvedValue({ byzantium: true, status: 1 }),
+      };
+      tokenContract.functions.approve.mockResolvedValueOnce(approveTx);
+
+      const setTotalDeposiTx: ContractTransaction = {
+        hash: txHash,
+        confirmations: 1,
+        nonce: 2,
+        gasLimit: bigNumberify(1e6),
+        gasPrice: bigNumberify(2e10),
+        value: Zero,
+        data: '0x',
+        chainId: depsMock.network.chainId,
+        from: depsMock.address,
+        wait: jest.fn().mockResolvedValue({ byzantium: true, status: 0 }),
+      };
+      tokenNetworkContract.functions.setTotalDeposit.mockResolvedValueOnce(setTotalDeposiTx);
+
+      await expect(
+        channelDepositEpic(action$, state$, depsMock).toPromise(),
+      ).resolves.toMatchObject({
+        type: getType(channelDepositFailed),
+        payload: expect.any(Error),
+        error: true,
+        meta: { tokenNetwork, partner },
+      });
+    });
+
+    test('success', async () => {
+      // there's a channel already opened in state
+      const curState = [
+        tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+        channelOpened(
+          { id: channelId, settleTimeout, openBlock, isFirstParticipant, txHash },
+          { tokenNetwork, partner },
+        ),
+        // own initial deposit of 330
+        channelDeposited(
+          {
+            id: channelId,
+            participant: depsMock.address,
+            totalDeposit: bigNumberify(330) as UInt<32>,
+            txHash,
+          },
+          { tokenNetwork, partner },
+        ),
+      ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(channelDeposit({ deposit }, { tokenNetwork, partner })),
+        state$ = of<RaidenState>(curState);
+
+      const approveTx: ContractTransaction = {
+        hash: txHash,
+        confirmations: 1,
+        nonce: 1,
+        gasLimit: bigNumberify(1e6),
+        gasPrice: bigNumberify(2e10),
+        value: Zero,
+        data: '0x',
+        chainId: depsMock.network.chainId,
+        from: depsMock.address,
+        wait: jest.fn().mockResolvedValue({ byzantium: true, status: 1 }),
+      };
+      tokenContract.functions.approve.mockResolvedValueOnce(approveTx);
+
+      const setTotalDepositTx: ContractTransaction = {
+        hash: txHash,
+        confirmations: 1,
+        nonce: 2,
+        gasLimit: bigNumberify(1e6),
+        gasPrice: bigNumberify(2e10),
+        value: Zero,
+        data: '0x',
+        chainId: depsMock.network.chainId,
+        from: depsMock.address,
+        wait: jest.fn().mockResolvedValue({ byzantium: true, status: 1 }),
+      };
+      tokenNetworkContract.functions.setTotalDeposit.mockResolvedValueOnce(setTotalDepositTx);
+
+      // result is undefined on success as the respective channelDepositedAction is emitted by the
+      // channelMonitoredEpic, which monitors the blockchain for ChannelNewDeposit events
+      await expect(
+        channelDepositEpic(action$, state$, depsMock).toPromise(),
+      ).resolves.toBeUndefined();
+      expect(tokenContract.functions.approve).toHaveBeenCalledTimes(1);
+      expect(approveTx.wait).toHaveBeenCalledTimes(1);
+      expect(tokenNetworkContract.functions.setTotalDeposit).toHaveBeenCalledTimes(1);
+      expect(tokenNetworkContract.functions.setTotalDeposit).toHaveBeenCalledWith(
+        channelId,
+        depsMock.address,
+        deposit.add(330),
+        partner,
+        expect.anything(),
+      );
+      expect(setTotalDepositTx.wait).toHaveBeenCalledTimes(1);
+    });
+  });
+
+  describe('channelCloseEpic', () => {
+    const openBlock = 121;
+
+    test('fails if there is no open channel with partner on tokenNetwork', async () => {
+      // there's a channel already opened in state
+      const action$ = of<RaidenAction>(channelClose(undefined, { tokenNetwork, partner })),
+        state$ = of<RaidenState>(state);
+
+      await expect(channelCloseEpic(action$, state$, depsMock).toPromise()).resolves.toMatchObject(
+        {
+          type: getType(channelCloseFailed),
+          payload: expect.any(Error),
+          error: true,
+          meta: { tokenNetwork, partner },
+        },
+      );
+    });
+
+    test('fails if channel.state !== "open"|"closing"', async () => {
+      // there's a channel already opened in state
+      const curState = [
+        tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+        // channel is in 'opening' state
+        channelOpen({ settleTimeout }, { tokenNetwork, partner }),
+      ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(channelClose(undefined, { tokenNetwork, partner })),
+        state$ = of<RaidenState>(curState);
+
+      await expect(channelCloseEpic(action$, state$, depsMock).toPromise()).resolves.toMatchObject(
+        {
+          type: getType(channelCloseFailed),
+          payload: expect.any(Error),
+          error: true,
+          meta: { tokenNetwork, partner },
+        },
+      );
+    });
+
+    test('closeChannel tx fails', async () => {
+      // there's a channel already opened in state
+      const curState = [
+        tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+        channelOpened(
+          { id: channelId, settleTimeout, openBlock, isFirstParticipant, txHash },
+          { tokenNetwork, partner },
+        ),
+      ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(channelClose(undefined, { tokenNetwork, partner })),
+        state$ = of<RaidenState>(curState);
+
+      const closeTx: ContractTransaction = {
+        hash: txHash,
+        confirmations: 1,
+        nonce: 2,
+        gasLimit: bigNumberify(1e6),
+        gasPrice: bigNumberify(2e10),
+        value: Zero,
+        data: '0x',
+        chainId: depsMock.network.chainId,
+        from: depsMock.address,
+        wait: jest.fn().mockResolvedValue({ byzantium: true, status: 0 }),
+      };
+      tokenNetworkContract.functions.closeChannel.mockResolvedValueOnce(closeTx);
+
+      await expect(channelCloseEpic(action$, state$, depsMock).toPromise()).resolves.toMatchObject(
+        {
+          type: getType(channelCloseFailed),
+          payload: expect.any(Error),
+          error: true,
+          meta: { tokenNetwork, partner },
+        },
+      );
+    });
+
+    test('success', async () => {
+      // there's a channel already opened in state
+      const curState = [
+        tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+        channelOpened(
+          { id: channelId, settleTimeout, openBlock, isFirstParticipant, txHash },
+          { tokenNetwork, partner },
+        ),
+      ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(channelClose(undefined, { tokenNetwork, partner })),
+        state$ = of<RaidenState>(curState);
+
+      const closeTx: ContractTransaction = {
+        hash: txHash,
+        confirmations: 1,
+        nonce: 3,
+        gasLimit: bigNumberify(1e6),
+        gasPrice: bigNumberify(2e10),
+        value: Zero,
+        data: '0x',
+        chainId: depsMock.network.chainId,
+        from: depsMock.address,
+        wait: jest.fn().mockResolvedValue({ byzantium: true, status: 1 }),
+      };
+      tokenNetworkContract.functions.closeChannel.mockResolvedValueOnce(closeTx);
+
+      // result is undefined on success as the respective channelClosedAction is emitted by the
+      // channelMonitoredEpic, which monitors the blockchain for channel events
+      await expect(
+        channelCloseEpic(action$, state$, depsMock).toPromise(),
+      ).resolves.toBeUndefined();
+      expect(tokenNetworkContract.functions.closeChannel).toHaveBeenCalledTimes(1);
+      expect(tokenNetworkContract.functions.closeChannel).toHaveBeenCalledWith(
+        channelId,
+        partner,
+        depsMock.address,
+        HashZero, // balance_hash
+        Zero, // nonce
+        HashZero, // additional_hash
+        expect.any(String), // non_closing_signature
+        expect.any(String), // closing_signature
+      );
+      expect(closeTx.wait).toHaveBeenCalledTimes(1);
+    });
+  });
+
+  describe('channelSettleEpic', () => {
+    const openBlock = 121,
+      closeBlock = 125,
+      settleBlock = closeBlock + settleTimeout + 1;
+
+    test('fails if there is no channel with partner on tokenNetwork', async () => {
+      // there's a channel already opened in state
+      const action$ = of<RaidenAction>(channelSettle(undefined, { tokenNetwork, partner })),
+        state$ = of<RaidenState>(state);
+
+      await expect(
+        channelSettleEpic(action$, state$, depsMock).toPromise(),
+      ).resolves.toMatchObject({
+        type: getType(channelSettleFailed),
+        payload: expect.any(Error),
+        error: true,
+        meta: { tokenNetwork, partner },
+      });
+    });
+
+    test('fails if channel.state !== "settleable|settling"', async () => {
+      // there's a channel in closed state, but not yet settleable
+      const curState = [
+        tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+        channelOpened(
+          { id: channelId, settleTimeout, openBlock, isFirstParticipant, txHash },
+          { tokenNetwork, partner },
+        ),
+        newBlock({ blockNumber: closeBlock }),
+        channelClosed(
+          { id: channelId, participant: depsMock.address, closeBlock, txHash },
+          { tokenNetwork, partner },
+        ),
+      ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(channelSettle(undefined, { tokenNetwork, partner })),
+        state$ = of<RaidenState>(curState);
+
+      await expect(
+        channelSettleEpic(action$, state$, depsMock).toPromise(),
+      ).resolves.toMatchObject({
+        type: getType(channelSettleFailed),
+        payload: expect.any(Error),
+        error: true,
+        meta: { tokenNetwork, partner },
+      });
+    });
+
+    test('settleChannel tx fails', async () => {
+      // there's a channel with partner in closed state and current block >= settleBlock
+      const curState = [
+        tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+        channelOpened(
+          { id: channelId, settleTimeout, openBlock, isFirstParticipant, txHash },
+          { tokenNetwork, partner },
+        ),
+        newBlock({ blockNumber: closeBlock }),
+        channelClosed(
+          { id: channelId, participant: depsMock.address, closeBlock, txHash },
+          { tokenNetwork, partner },
+        ),
+        newBlock({ blockNumber: settleBlock }),
+        channelSettleable({ settleableBlock: settleBlock }, { tokenNetwork, partner }),
+      ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(channelSettle(undefined, { tokenNetwork, partner })),
+        state$ = of<RaidenState>(curState);
+
+      const settleTx: ContractTransaction = {
+        hash: txHash,
+        confirmations: 1,
+        nonce: 2,
+        gasLimit: bigNumberify(1e6),
+        gasPrice: bigNumberify(2e10),
+        value: Zero,
+        data: '0x',
+        chainId: depsMock.network.chainId,
+        from: depsMock.address,
+        wait: jest.fn().mockResolvedValue({ byzantium: true, status: 0 }),
+      };
+      tokenNetworkContract.functions.settleChannel.mockResolvedValueOnce(settleTx);
+
+      await expect(
+        channelSettleEpic(action$, state$, depsMock).toPromise(),
+      ).resolves.toMatchObject({
+        type: getType(channelSettleFailed),
+        payload: expect.any(Error),
+        error: true,
+        meta: { tokenNetwork, partner },
+      });
+    });
+
+    test('success', async () => {
+      // there's a channel with partner in closed state and current block >= settleBlock
+      const curState = [
+        tokenMonitored({ token, tokenNetwork, fromBlock: 1 }),
+        channelOpened(
+          { id: channelId, settleTimeout, openBlock, isFirstParticipant, txHash },
+          { tokenNetwork, partner },
+        ),
+        newBlock({ blockNumber: closeBlock }),
+        channelClosed(
+          { id: channelId, participant: depsMock.address, closeBlock, txHash },
+          { tokenNetwork, partner },
+        ),
+        newBlock({ blockNumber: settleBlock }),
+        channelSettleable({ settleableBlock: settleBlock }, { tokenNetwork, partner }),
+      ].reduce(raidenReducer, state);
+      const action$ = of<RaidenAction>(channelSettle(undefined, { tokenNetwork, partner })),
+        state$ = of<RaidenState>(curState);
+
+      const settleTx: ContractTransaction = {
+        hash: txHash,
+        confirmations: 1,
+        nonce: 2,
+        gasLimit: bigNumberify(1e6),
+        gasPrice: bigNumberify(2e10),
+        value: Zero,
+        data: '0x',
+        chainId: depsMock.network.chainId,
+        from: depsMock.address,
+        wait: jest.fn().mockResolvedValue({ byzantium: true, status: 1 }),
+      };
+      tokenNetworkContract.functions.settleChannel.mockResolvedValueOnce(settleTx);
+
+      // result is undefined on success as the respective ChannelSettledAction is emitted by the
+      // channelMonitoredEpic, which monitors the blockchain for channel events
+      await expect(
+        channelSettleEpic(action$, state$, depsMock).toPromise(),
+      ).resolves.toBeUndefined();
+      expect(tokenNetworkContract.functions.settleChannel).toHaveBeenCalledTimes(1);
+      expect(tokenNetworkContract.functions.settleChannel).toHaveBeenCalledWith(
+        channelId,
+        depsMock.address,
+        expect.anything(), // self transfered amount
+        expect.anything(), // self locked amount
+        expect.anything(), // self locksroot
+        partner,
+        expect.anything(), // partner transfered amount
+        expect.anything(), // partner locked amount
+        expect.anything(), // partner locksroot
+      );
+      expect(settleTx.wait).toHaveBeenCalledTimes(1);
+    });
+  });
+});
