+++ raiden-ts/tests/unit/mocks.ts
@@ -1,39 +1,8 @@
 /* eslint-disable @typescript-eslint/no-explicit-any,@typescript-eslint/camelcase */
 
-// ethers's contracts use a lot defineReadOnly which doesn't allow us to mock
-// functions and properties. Mock it here so we can mock later
-jest.mock('ethers/utils/properties', () => ({
-  ...jest.requireActual('ethers/utils/properties'),
-  defineReadOnly: jest.fn(
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (object: any, name: string, value: any): void =>
-      Object.defineProperty(object, name, {
-        enumerable: true,
-        value,
-        writable: true,
-        configurable: true,
-      }),
-  ),
-}));
-
-// ethers utils mock to always validate matrix userIds/displayName
-jest.mock('ethers/utils', () => ({
-  ...jest.requireActual('ethers/utils'),
-  verifyMessage: jest.fn((msg: string, sig: string): string => {
-    const { getAddress, verifyMessage: origVerifyMessage } = jest.requireActual('ethers/utils');
-    const match = /^@(0x[0-9a-f]{40})[.:]/i.exec(msg);
-    if (match && match[1]) return getAddress(match[1]);
-    return origVerifyMessage(msg, sig);
-  }),
-}));
-
-// raiden-ts/utils.getNetwork has the same functionality as provider.getNetwork
-// but fetches everytime instead of just returning a cached property
-// On mocked tests, we unify both again, so we can just mock provider.getNetwork in-place
-jest.mock('raiden-ts/utils/matrix', () => ({
-  ...jest.requireActual('raiden-ts/utils/matrix'),
-  getNetwork: jest.fn((provider: JsonRpcProvider): Promise<Network> => provider.getNetwork()),
-}));
+import { patchEthersDefineReadOnly, patchMatrixGetNetwork } from './patches';
+patchEthersDefineReadOnly();
+patchMatrixGetNetwork();
 
 import { BehaviorSubject, Subject, AsyncSubject, of } from 'rxjs';
 import { MatrixClient } from 'matrix-js-sdk';
@@ -60,7 +29,7 @@ import { RaidenState, initialState } from 'raiden-ts/state';
 import { Address, Signature } from 'raiden-ts/utils/types';
 import { RaidenConfig, defaultConfig } from 'raiden-ts/config';
 
-type MockedContract<T extends Contract> = jest.Mocked<T> & {
+export type MockedContract<T extends Contract> = jest.Mocked<T> & {
   functions: {
     [K in keyof T['functions']]: jest.MockInstance<
       ReturnType<T['functions'][K]>,
@@ -69,7 +38,7 @@ type MockedContract<T extends Contract> = jest.Mocked<T> & {
   };
 };
 
-interface MockRaidenEpicDeps extends RaidenEpicDeps {
+export interface MockRaidenEpicDeps extends RaidenEpicDeps {
   provider: jest.Mocked<JsonRpcProvider>;
   registryContract: MockedContract<TokenNetworkRegistry>;
   getTokenNetworkContract: (address: string) => MockedContract<TokenNetwork>;
